+++ packages/react-form-schema-ui-antd/__stories__/index.stories.tsx
@@ -1,19 +1,19 @@
-import type { Meta, StoryObj } from '@storybook/react';
+import type { Meta, StoryObj } from "@storybook/react";
 
-import { Simple } from '../src/examples/Simple';
+import { Simple } from "../src/examples/Simple";
 
 const meta: Meta = {
-  title: 'react-form-schema-ui-antd',
-  tags: ['autodocs'],
+	title: "react-form-schema-ui-antd",
+	tags: ["autodocs"],
 };
 
 export default meta;
 type Story = StoryObj;
 
 export const SimpleStory: Story = {
-  name: 'Simple',
-  args: {
-    required: false,
-  },
-  render: (props) => <Simple {...props} />,
+	name: "Simple",
+	args: {
+		required: false,
+	},
+	render: (props) => <Simple {...props} />,
 };