+++ packages/react-hook-form-schema/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added built-in field `list`
 - Added `fieldPath` render parameter
 - Added `renderPath` param to `set` field payload
+- Added `excludePaths` param to `set` field payload
 
 ## 1.1.0 (26 nov 2023)